+++ dakara_server/library/serializers.py
@@ -123,8 +123,10 @@ class SongTagSerializer(serializers.ModelSerializer):
     class Meta:
         model = SongTag
         fields = (
+                'id',
                 'name',
                 'color_id',
+                'disabled',
                 )