+++ dakara_server/library/tests_song.py
@@ -2,12 +2,17 @@ from django.core.urlresolvers import reverse
 from rest_framework import status
 from .base_test import BaseAPITestCase
 
+from .models import SongTag
+
 class SongListViewAPIViewTestCase(BaseAPITestCase):
     url = reverse('library-song-list')
 
     def setUp(self):
         # create a user without any rights
         self.user = self.create_user("TestUser")
+
+        # create a manager
+        self.manager = self.create_user("TestManager", library_level='m')
 
         # create test data
         self.create_library_test_data()
@@ -197,3 +202,47 @@ class SongListViewAPIViewTestCase(BaseAPITestCase):
         self.assertEqual(len(results), len(expected_songs))
         for song, expected_song in zip(results, expected_songs):
             self.assertEqual(song['id'], expected_song.id)
+
+    def test_get_song_list_disabled_tag(self):
+        """
+        Test to verify that song list does not include disabled songs with tags
+        """
+        # Login as simple user
+        self.authenticate(self.user)
+
+        # Set tag1 disabled
+        self.tag1.disabled = True
+        self.tag1.save()
+
+        # Get songs list, there should be only one song
+        response = self.client.get(self.url)
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data['count'], 1)
+        self.assertEqual(len(response.data['results']), 1)
+
+        # The remaining song is Song1
+        # Song2 has been excluded from the list
+        self.check_song_json(response.data['results'][0], self.song1)
+
+    def test_get_song_list_disabled_tag_manager(self):
+        """
+        Test to verify that song list includes disabled songs with tags if the
+        user is a manager
+        """
+        # Login as simple user
+        self.authenticate(self.manager)
+
+        # Set tag1 disabled
+        self.tag1.disabled = True
+        self.tag1.save()
+
+        # Get songs list, there should be only one song
+        response = self.client.get(self.url)
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data['count'], 2)
+        self.assertEqual(len(response.data['results']), 2)
+
+        # The remaining song is Song1
+        # Song2 has been excluded from the list
+        self.check_song_json(response.data['results'][0], self.song1)
+        self.check_song_json(response.data['results'][1], self.song2)