+++ CHANGELOG.md
@@ -33,9 +33,11 @@
 ### Added
 - New option for createtags and createworktypes commands: --prune, to remove tags and worktypes not in config file.
 - New Command: prune, to remove works and artists no longer linked to a song.
+- Allow to disable tags: song with disabled tags are visible only by the library manager and can be added to the playlist only by the playlist manager.
 
 ### Fixed
-- When the feeder update a song, the song is no longer associated with old artists, works and tags
+- When the feeder update a song, the song is no longer associated with old artists, works and tags.
+- Only library manager can update library.
 
 ## 1.0.0 - 2017-11-12